% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brasileirao.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Encontrar o pior ano de um time}
\usage{
encontrar_pior_ano_time(time, tirar_ano_atual = TRUE)
}
\arguments{
\item{time}{Uma string contendo o time.}

\item{tirar_ano_atual}{Tira o ano atual.}
}
\value{
Um data frame com 3 colunas e 1 linha.
}
\description{
A funcao parte da base \link{partidas_brasileirao} e
encontra o pior ano em numero de vitorias para
um dado time. Olhe tambem \link{sumarizar_paises}
}
\examples{
# Encontrar o pior ano do Flamengo
encontrar_pior_ano_time("Flamengo")
}
\seealso{
A base \link{partidas_brasileirao}.
}
