% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{europa}
\alias{europa}
\title{Europa}
\format{
uma tibble com 2 linhas e 3 colunas
\describe{
\item{ano}{ano de referência}
\item{pop}{população no continente para o ano de referência}
\item{vida}{expectativa de vida para o ano de referência}
}
}
\source{
\url{https://cienciadedatos.github.io/dados/reference/dados_gapminder.html}
}
\usage{
europa
}
\description{
Base de exemplo, com população e expectativa de vida no continente
europeu nos anos 2002 e 2007.
}
\seealso{
\link[dados:dados_gapminder]{dados::dados_gapminder}
}
\keyword{datasets}
