% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumarizar_paises.R
\name{sumarizar_paises}
\alias{sumarizar_paises}
\title{Title}
\usage{
sumarizar_paises(continentes = "Ásia", ano_minimo = 2000)
}
\arguments{
\item{continentes}{bla}

\item{ano_minimo}{bla}
}
\value{
bla
}
\description{
Title
}
