% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Partidas do Brasileirão}
\format{
Uma base com 8026 linhas e 8 colunas:
\describe{
\item{temporada}{Temporada do jogo (numeric)}
\item{data}{Data do jogo (date)}
\item{time_casa}{Nome do time da casa}
\item{time_visitante}{Nome do time visitante}
}
}
\source{
Site do Brasileirão \url{https://github.com/williamorim/brasileirao}
}
\usage{
partidas_brasileirao
}
\description{
Uma base de dados que contém blablabla...
}
\keyword{datasets}
